changes in the indexer and rest server
export the entire set of documents as json and run the indexer to get the tfidf file and import the file to mongo after deleting the original collection and then start the server. note- this is for online mongodb, make appropriate changes for the local db in the pymongotest file.